Rewriting PDB reading to handle PDB and PDBx/mmCIF - Everything is broken from here
